+++ Sources/PippinLibrary/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - Different ways to count things in Collections: `count(where:)` and `counts`.
 - Ability to print a 2D collection with `gridDescription`.
+- Extract typed values from strings like `lines`, `ints`, `stringsAndInts`, `intGrid` and `characterGrid`.
 
 ## [2.6.0] 2020-10-24